test: shared structure refactoring
